+++ packages/css-calc/src/index.ts
@@ -32,3 +32,28 @@ export function calcFromComponentValues(componentValuesList: Array<Array<Compone
 		}
 	});
 }
+
+// Exposed so upstream dependents can use the same set of math functions.
+export const mathFunctionNames = new Set([
+	'abs',
+	'acos',
+	'asin',
+	'atan',
+	'atan2',
+	'calc',
+	'clamp',
+	'cos',
+	'exp',
+	'hypot',
+	'log',
+	'max',
+	'min',
+	'mod',
+	'pow',
+	'rem',
+	'round',
+	'sign',
+	'sin',
+	'sqrt',
+	'tan',
+]);