+++ packages/css-color-parser/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to CSS Color Parser
 
+### Unreleased (patch)
+
+- Improve the detection of math function in color notations.
+
 ### 1.1.0 (March 28, 2023)
 
 - Add a flag to `serializeP3` and `serializeRGB` to skip gamut mapping.